Added tests for TgWebHookHandler

+++ lib/bodhi/tg_webhook_handler.ex
@@ -10,11 +10,16 @@ defmodule Bodhi.TgWebhookHandler do
   alias Telegex.Type.{Message, Update, MessageEntity}
   alias Bodhi.Prompts.Prompt
 
+  @user_commands ~w(start)
+
   @impl true
   def on_boot() do
     # env_config = Application.get_env(:bodhi, __MODULE__)
     # delete the webhook and set it again
-    {:ok, true} = Telegex.delete_webhook()
+    unless Mix.env() == :test do
+      {:ok, true} = Telegex.delete_webhook()
+    end
+
     # {:ok, bot_user} = Telegex.get_me()
     # Bodhi.Users.create_or_update_user(bot_user)
     # set the webhook (url is required)
@@ -113,6 +118,10 @@ defmodule Bodhi.TgWebhookHandler do
     )
 
     {:ok, answer}
+  end
+
+  defp get_answer(%_{chat_id: chat_id, text: "/" <> _}, lang) do
+    {:ok, "Unknowwn command. Please use /start to begin."}
   end
 
   defp get_answer(%_{chat_id: chat_id}, _) do
